# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Imposter AI API
  version: 1.0.0
  description: |
    # Imposter AI — Backend Specification

    Backend API для игры Imposter AI — управление категориями, AI-генерацией слов, premium-контентом и remote config.

    ---

    ## Сущности

    ### Category
    Описывает тематическую категорию слов. Делятся на free и premium через флаг `is_premium`.
    Активность управляется через `is_active` — скрывает без удаления. `sort_order` определяет порядок в клиенте.
    Все текстовые поля хранятся как LocalizedString — JSON-объект с ключами по ISO locale.

    ### WordPack
    Набор слов привязанный к категории. Создаётся отдельно под каждую локаль — автоматического перевода нет.
    - Для стандартного режима `impostor_word = null` — impostor не получает слово.
    - Для Party Mode (Undercover) `impostor_word` содержит похожее но отличное слово.
    - При запросе клиент передаёт `locale`, сервер ищет подходящий пак; если нет — fallback на `en`.

    ### FeatureConfig
    Управление feature flags и контент-гейтингом. Клиент получает весь конфиг одним запросом `GET /config` при старте.
    Флаги кешируются на клиенте с TTL 15–30 мин. `payload` может содержать лимиты AI-запросов для free-tier.
    Ключи: `"ai_theme"`, `"premium_categories"`, `"party_mode_extra_roles"`.

    ### AITopicRequestLog
    Лог запросов на AI-генерацию — для дебага, модерации, rate limiting. Хранится для аудита, не возвращается клиенту.
...

    ### LocalizationContent
...

    ---

...

...

...

...

...

...

...

...
servers:
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Api-Key
  parameters:
    DeviceId:
      name: X-Device-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Анонимный идентификатор устройства, генерируется при первом запуске
    AppVersion:
      name: X-App-Version
      in: header
      required: true
      schema:
        type: string
        example: "1.0.0"
    LocaleHeader:
      name: X-Locale
      in: header
      required: false
      schema:
        type: string
        example: "ru"
      description: ISO 639-1 код языка
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: RATE_LIMIT_EXCEEDED
            message:
              type: string
              example: AI request limit reached for today
            details:
              nullable: true
              example: null
...
    AIThemeResponse:
      type: object
      properties:
        locale:
          type: string
          example: ru
        topic:
          type: string
          example: космос
        civilian_word:
          type: string
          example: Марс
        impostor_word:
          type: string
          nullable: true
          example: null
        difficulty:
          type: string
          enum: [easy, medium, hard]
        is_safe:
          type: boolean
        fallback_used:
          type: boolean
        fallback_reason:
          type: string
          nullable: true
          enum: [ai_unavailable, moderation_rejected, rate_limit_exceeded]
paths:
...
  /ai/generate-theme:
    post:
      summary: Генерация слова по теме
      description: |
        Генерирует слово или пару слов по пользовательской теме через LLM.

        **Логика пула слов:** сервер хранит историю выданных слов по device_id с TTL 30 дней.
        При повторном запросе той же темы выдаётся слово которое устройство ещё не получало.
        Если все слова исчерпаны — LLM генерирует новые и добавляет в пул.

        **Fallback:** при недоступности LLM, превышении rate limit или отклонении модерацией
        возвращается curated слово с fallback_used: true. Клиент обрабатывает как успех.
      tags: [AI]
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/AppVersion'
        - $ref: '#/components/parameters/LocaleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic, locale]
              properties:
                topic:
                  type: string
                  maxLength: 80
                  example: космос
                  description: Пользовательская тема
                locale:
                  type: string
                  example: ru
                  description: ISO locale — язык ответа
                mode:
                  type: string
                  enum: [standard, party]
                  default: standard
                  description: standard — одно слово для civilians; party — civilian_word + impostor_word
      responses:
        '200':
          description: |
            Успешный ответ. Включает fallback-случаи (fallback_used: true).
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              example: 5
            X-RateLimit-Remaining:
              schema:
                type: integer
              example: 2
            X-RateLimit-Reset:
              schema:
                type: integer
              description: Unix timestamp когда освободится первый слот
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIThemeResponse'
              examples:
                standard:
                  summary: Стандартный режим
                  value:
                    locale: ru
                    topic: космос
                    civilian_word: Марс
                    impostor_word: null
                    difficulty: medium
                    is_safe: true
                    fallback_used: false
                party:
                  summary: Party mode
                  value:
                    locale: ru
                    topic: космос
                    civilian_word: Марс
                    impostor_word: Луна
                    difficulty: medium
                    is_safe: true
                    fallback_used: false
                fallback:
                  summary: Fallback ответ
                  value:
                    locale: ru
                    topic: космос
                    civilian_word: Земля
                    impostor_word: null
                    difficulty: easy
                    is_safe: true
                    fallback_used: true
                    fallback_reason: ai_unavailable
        '400':
          description: topic пустой или превышает 80 символов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: |
            Превышен rate limit. Тело содержит fallback_used: true + curated слово.
            Клиент обрабатывает как успех, не как ошибку.
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: 0
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIThemeResponse'
              example:
                locale: ru
                topic: космос
                civilian_word: Земля
                impostor_word: null
                difficulty: easy
                is_safe: true
                fallback_used: true
                fallback_reason: rate_limit_exceeded
        '500':
          description: Внутренняя ошибка (только если нет контента для fallback)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: LLM или внешний сервис временно недоступен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /ai/generate-words:
    post:
      summary: Генерация набора слов для WordPack (только admin)
      description: |
        Генерирует полный набор слов по теме для наполнения WordPack.
        Только для internal use — admin-панели и контент-команды.
        Недоступен обычным пользователям (free и premium).
      tags: [AI]
      security:
        - AdminApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic, locale]
              properties:
                topic:
                  type: string
                  example: Space exploration
                locale:
                  type: string
                  example: en
                count:
                  type: integer
                  minimum: 10
                  maximum: 30
                  default: 20
                  example: 20
                  description: |
                    Желаемое количество слов. Минимум 10, по умолчанию 20, максимум 30.
                    Если LLM вернул меньше 10 слов — сервер автоматически повторяет запрос
                    и добирает слова до минимума без участия клиента.
                mode:
                  type: string
                  enum: [standard, party]
                  default: standard
                difficulty:
                  type: string
                  enum: [easy, medium, hard]
                  example: medium
      responses:
        '200':
          description: Сгенерированный набор слов
          content:
            application/json:
              schema:
                type: object
                properties:
                  locale:
                    type: string
                  topic:
                    type: string
                  words:
                    type: array
                    items:
                      type: object
                      properties:
                        civilian_word:
                          type: string
                        impostor_word:
                          type: string
                          nullable: true
                        difficulty:
                          type: string
                          enum: [easy, medium, hard]
                        is_safe:
                          type: boolean
                  total:
                    type: integer
                    description: Количество возвращённых слов после фильтрации
                  unsafe_filtered:
                    type: integer
                    description: Количество слов отфильтрованных модерацией
              example:
                locale: en
                topic: Space exploration
                words:
                  - civilian_word: Mars
                    impostor_word: Moon
                    difficulty: medium
                    is_safe: true
                  - civilian_word: Rocket
                    impostor_word: Satellite
                    difficulty: easy
                    is_safe: true
                total: 19
                unsafe_filtered: 1
        '400':
          description: Невалидные поля запроса (topic или locale отсутствуют)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Отсутствует или невалидный X-Admin-Api-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: LLM временно недоступен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...